+++ tests/test_commands.py
@@ -4,6 +4,9 @@ from tempfile import TemporaryDirectory
 import pystac
 from stactools.nrcan_radarsat1.commands import create_nrcanradarsat1_command
 from stactools.testing import CliTestCase
+from stactools.testing import TestData
+
+test_data = TestData(__file__)
 
 
 class CommandsTest(CliTestCase):
@@ -12,13 +15,8 @@ class CommandsTest(CliTestCase):
 
     def test_create_collection(self):
         with TemporaryDirectory() as tmp_dir:
-            # Run your custom create-collection command and validate
-
-            # Example:
-            destination = os.path.join(tmp_dir, "collection.json")
-
             result = self.run_command(
-                ["nrcanradarsat1", "create-collection", destination])
+                ["nrcanradarsat1", "create-collection", "-d", tmp_dir])
 
             self.assertEqual(result.exit_code,
                              0,
@@ -27,23 +25,28 @@ class CommandsTest(CliTestCase):
             jsons = [p for p in os.listdir(tmp_dir) if p.endswith(".json")]
             self.assertEqual(len(jsons), 1)
 
-            collection = pystac.read_file(destination)
-            self.assertEqual(collection.id, "my-collection-id")
+            collection = pystac.read_file(os.path.join(tmp_dir, jsons[0]))
+            # self.assertEqual(collection.id, "my-collection-id")
             # self.assertEqual(item.other_attr...
 
             collection.validate()
 
     def test_create_item(self):
         with TemporaryDirectory() as tmp_dir:
-            # Run your custom create-item command and validate
 
-            # Example:
-            destination = os.path.join(tmp_dir, "collection.json")
+            test_path = test_data.get_path("data-files")
+            # Select a .tif data file
+            cog_path = os.path.join(test_path, [
+                d for d in os.listdir(test_path) if d.lower().endswith(".tif")
+            ][0])
+
             result = self.run_command([
                 "nrcanradarsat1",
                 "create-item",
-                "/path/to/asset.tif",
-                destination,
+                "-s",
+                cog_path,
+                "-d",
+                tmp_dir,
             ])
             self.assertEqual(result.exit_code,
                              0,
@@ -52,8 +55,36 @@ class CommandsTest(CliTestCase):
             jsons = [p for p in os.listdir(tmp_dir) if p.endswith(".json")]
             self.assertEqual(len(jsons), 1)
 
-            item = pystac.read_file(destination)
-            self.assertEqual(item.id, "my-item-id")
-            # self.assertEqual(item.other_attr...
+            item_path = os.path.join(tmp_dir, jsons[0])
+            item = pystac.read_file(item_path)
 
             item.validate()
+
+    # Downloads full cog file. Suggest leaving commented unless desired to test
+    def test_download_asset(self):
+        enabled = False
+        if enabled:
+            with TemporaryDirectory() as tmp_dir:
+                cog_href = (
+                    "s3://radarsat-r1-l1-cog/2012"
+                    "/8/RS1_B0625465_SCWA_20120822_122459_HH_SCW01F.tif")
+
+            result = self.run_command([
+                "nrcanradarsat1",
+                "download-asset",
+                "-s",
+                cog_href,
+                "-d",
+                tmp_dir,
+            ])
+
+            self.assertEqual(result.exit_code,
+                             0,
+                             msg="\n{}".format(result.output))
+
+            cogs = [p for p in os.listdir(tmp_dir) if p.endswith(".tif")]
+            self.assertEqual(len(cogs), 1)
+        else:
+            print(
+                "\nCog download test disabled. Re-enable by setting enabled = True "
+                "on line 65 in tests/test_commands.py\n")